Refactor client to conform to AirBnB style
