+++ __tests__/client/BidInfo.enzyme.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import TopInfo from '../../client/TopInfo.js';
+import TopInfo from '../../client/TopInfo';
 import { shallow, configure } from 'enzyme';
 import placeholderProduct from './placeholderProduct';
 import Adapter from 'enzyme-adapter-react-16';
