+++ __tests__/client/TopInfo.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import TopInfo from '../../client/TopInfo.js';
+import TopInfo from '../../client/TopInfo';
 import renderer from 'react-test-renderer';
-import placeholderProduct from './placeholderProduct.js';
+import placeholderProduct from './placeholderProduct';
 
 it('renders correctly', () => {
   const tree = renderer
