+++ __tests__/client/placeholderProduct.js
