+++ client/BidInfo.jsx
