--- client/index.js
@@ -1,5 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import BidBuy from './BidBuy.js';
-
-ReactDOM.render(<BidBuy />, document.getElementById('app'));
