+++ client/index.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import BidBuy from './BidBuy';
+
+ReactDOM.render(<BidBuy />, document.getElementById('app'));
