+++ package.json
@@ -32,6 +32,7 @@
     "webpack-cli": "^3.3.5"
   },
   "devDependencies": {
+    "babel-eslint": "^10.0.2",
     "babel-jest": "^24.8.0",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
