+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  entry: './client/index.js',
+  entry: './client/index.jsx',
   output: {
     filename: 'bundle.js',
     path: path.resolve(__dirname, 'public')
@@ -10,6 +10,9 @@ module.exports = {
     rules: [
       {
         test: /\.jsx?$/,
+        resolve: {
+          extensions: ['.js', '.jsx'],
+        },
         exclude: /node_modules/,
         use: {
           loader: 'babel-loader',
@@ -19,5 +22,5 @@ module.exports = {
         }
       }
     ]
-  }
+  },
 };
